requirements.txt
google-api-python-client==2.110.0
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
openai==1.6.1
matplotlib==3.8.2
seaborn==0.13.0
nltk==3.8.1
wordcloud==1.9.2
beautifulsoup4==4.12.2
youtube-transcript-api==0.6.1